Replaced some u8 and u16 by u32.

+++ source/Game.c
@@ -486,20 +486,20 @@ static void Game_handleInput(Game this)
 	// poll the user's input
 	KeypadManager_read(this->keypadManager);
 
-	u16 pressedKey = KeypadManager_getPressedKey(this->keypadManager);
-	u16 releasedKey = KeypadManager_getReleasedKey(this->keypadManager);
-	u16 holdKey = KeypadManager_getHoldKey(this->keypadManager);
+	u32 pressedKey = KeypadManager_getPressedKey(this->keypadManager);
+	u32 releasedKey = KeypadManager_getReleasedKey(this->keypadManager);
+	u32 holdKey = KeypadManager_getHoldKey(this->keypadManager);
 
 #ifdef __DEBUG_TOOLS
-	u16 previousKey = KeypadManager_getPreviousKey(this->keypadManager);
+	u32 previousKey = KeypadManager_getPreviousKey(this->keypadManager);
 #else
 
 #ifdef __STAGE_EDITOR
-	u16 previousKey = KeypadManager_getPreviousKey(this->keypadManager);
+	u32 previousKey = KeypadManager_getPreviousKey(this->keypadManager);
 #else
 
 #ifdef __ANIMATION_EDITOR
-	u16 previousKey = KeypadManager_getPreviousKey(this->keypadManager);
+	u32 previousKey = KeypadManager_getPreviousKey(this->keypadManager);
 #endif
 
 #endif
