+++ source/graphics/2d/Textures/Bgmap/BgmapTexture.h
@@ -76,9 +76,9 @@ __CLASS_NEW_DECLARE(BgmapTexture, BgmapTextureDefinition* bgmapTextureDefinition
 
 void BgmapTexture_destructor(BgmapTexture this);
 void BgmapTexture_write(BgmapTexture this);
-s8 BgmapTexture_getXOffset(BgmapTexture this);
-s8 BgmapTexture_getYOffset(BgmapTexture this);
-u8 BgmapTexture_getBgmapSegment(BgmapTexture this);
+s32 BgmapTexture_getXOffset(BgmapTexture this);
+s32 BgmapTexture_getYOffset(BgmapTexture this);
+u32 BgmapTexture_getBgmapSegment(BgmapTexture this);
 void BgmapTexture_increaseUsageCount(BgmapTexture this);
 bool BgmapTexture_decreaseUsageCount(BgmapTexture this);
