+++ source/hardware/KeypadManager.h
@@ -93,10 +93,10 @@ int KeypadManager_isEnabled(KeypadManager this);
 void KeypadManager_read(KeypadManager this);
 void KeypadManager_clear(KeypadManager this);
 void KeypadManager_flush(KeypadManager this);
-u16 KeypadManager_getPressedKey(KeypadManager this);
-u16 KeypadManager_getReleasedKey(KeypadManager this);
-u16 KeypadManager_getHoldKey(KeypadManager this);
-u16 KeypadManager_getPreviousKey(KeypadManager this);
+u32 KeypadManager_getPressedKey(KeypadManager this);
+u32 KeypadManager_getReleasedKey(KeypadManager this);
+u32 KeypadManager_getHoldKey(KeypadManager this);
+u32 KeypadManager_getPreviousKey(KeypadManager this);
 
 
 #endif
